requirements.txt
langchain
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
python-dotenv
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu 
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
duckduckgo-search
